+++ core/src/script/CGXP/plugins/FeaturesGrid.js
@@ -16,6 +16,7 @@
  */
 
 /*
+ * @requires CGXP/plugins/FeaturesResult.js
  * @include GeoExt/widgets/tree/LayerContainer.js
  * @include GeoExt/widgets/tree/LayerLoader.js
  * @include GeoExt/plugins/TreeNodeComponent.js
@@ -33,13 +34,13 @@
 
 /** api: (define)
  *  module = cgxp.plugins
- *  class = FeatureGrid
+ *  class = FeaturesGrid
  */
 
 Ext.namespace("cgxp.plugins");
 
 /** api: example
- *  Sample code showing how to add a FeatureGrid plugin to a
+ *  Sample code showing how to add a FeaturesGrid plugin to a
  *  `gxp.Viewer`:
  *
  *  .. code-block:: javascript
@@ -50,11 +51,12 @@ Ext.namespace("cgxp.plugins");
  *      new gxp.Viewer({
  *          ...
  *          tools: [{
- *              ptype: "cgxp_featuregrid",
- *              id: "featureGrid",
+ *              ptype: "cgxp_featuresgrid",
+ *              id: "featuresGrid",
  *              csvURL: "${request.route_url('csvecho')}",
  *              maxFeatures: 200,
- *              outputTarget: "featuregrid-container",
+ *              outputTarget: "featuresgrid-container",
+ *              themes: THEMES,
  *              events: EVENTS
  *          }, {
  *              ptype: "cgxp_wmsgetfeatureinfo",
@@ -67,7 +69,7 @@ Ext.namespace("cgxp.plugins");
  */
 
 /** api: constructor
- *  .. class:: FeatureGrid(config)
+ *  .. class:: FeaturesGrid(config)
  *
  *      A plugin that adds a grid panel for displaying feature information (one
  *      feature per row).
@@ -76,11 +78,25 @@ Ext.namespace("cgxp.plugins");
  *      :class:`cgxp.plugins.WMSGetFeatureInfo`, and
  *      :class:`cgxp.plugins.QueryBuilder`.
  *
+ *      For the queryable Base layer the ``identifierAttribute`` can be
+ *      provide by the ``queryLayers`` option in a layer config:
+ *
+ *      .. code-block:: javascript
+ *
+ *          ...
+ *          queryLayers: [{
+ *              name: "buildings",
+ *              identifierAttribute: "name"
+ *          }, {
+ *              name: "parcels",
+ *              identifierAttribute: "number"
+ *          }]
+ *          ...
  */
-cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
+cgxp.plugins.FeaturesGrid = Ext.extend(cgxp.plugins.FeaturesResult, {
 
-    /** api: ptype = cgxp_featuregrid */
-    ptype: "cgxp_featuregrid",
+    /** api: ptype = cgxp_featuresgrid */
+    ptype: "cgxp_featuresgrid",
 
     /** private: attibute[this.tabpan]
      *  ``Object``
@@ -92,12 +108,6 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
      *  The visible grid.
      */
     currentGrid: null,
-
-    /** private: attibute[vectorLayer]
-     *  ``OpenLayers.Layer.Vector``
-     *  The layer used to display the features.
-     */
-    vectorLayer: null,
 
     /** private: attibute[gridByType]
      *  ``Object``
@@ -139,6 +149,12 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
      *  * ``queryresults(features)``: sent when the result is received
      */
     events: null,
+
+    /** api: config[themes]
+     *  ``Object`` List of internal and external themes and layers. (The
+     *  same object as passed to the :class:`cgxp.plugins.LayerTree`).
+     */
+    themes: null,
 
     /** api: config[globalSelection]
      *  ``Boolean`` If true, selection state are remembered across all result
@@ -226,7 +242,7 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
      */
     init: function() {
         this.dummyForm = Ext.DomHelper.append(document.body, {tag : 'form'});
-        cgxp.plugins.FeatureGrid.superclass.init.apply(this, arguments);
+        cgxp.plugins.FeaturesGrid.superclass.init.apply(this, arguments);
         this.target.on('ready', this.viewerReady, this);
     },
 
@@ -366,10 +382,9 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
      *  Set the queryResult message, check if there is enough space to display it all
      */
     setMessage: function(msg) {
-        var msg = msg;
         // tests the space required by the TextItem
         this.messageItem.setText(msg);
-        
+
         if ((this.tabpan.getInnerWidth() - 370) < this.messageItem.getWidth()) {
             msg = [
                 '<abbr title="',
@@ -400,8 +415,7 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
      *  ``Ext.data.Record``
      */
     showFeature: function(record) {
-        record.getFeature().style = OpenLayers.Feature.Vector.style['default'];
-        record.getFeature().style.strokeWidth = 4;
+        record.getFeature().style = null;
         this.vectorLayer.drawFeature(record.getFeature());
     },
 
@@ -444,10 +458,12 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
 
         // a ResultsPanel object has its own vector layer, which
         // is added to the map once for good
-        this.vectorLayer = new OpenLayers.Layer.Vector(
-            OpenLayers.Util.createUniqueID("c2cgeoportal"), {
-                displayInLayerSwitcher: false,
-                alwaysInRange: true
+        this.createVectorLayer({
+            styleMap: new OpenLayers.StyleMap({
+                'select': {
+                    'strokeWidth': 4
+                }
+            })
         });
 
         this.events.on('querystarts', function() {
@@ -483,7 +499,9 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
         }, this);
 
         this.events.on('queryclose', function() {
-            this.control && this.control.deactivate();
+            if (this.control) {
+                this.control.deactivate();
+            }
         }, this);
 
         this.events.on('queryresults', function(queryResult, selectAll) {
@@ -494,7 +512,7 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
 
             // if no feature do nothing
             if ((!features || features.length === 0) &&
-                    (!queryResult.unqueriedLayers || 
+                    (!queryResult.unqueriedLayers ||
                     queryResult.unqueriedLayers.length === 0)) {
                 // if really no feature close panel)
                 if (previouslyNoFeature) {
@@ -791,8 +809,8 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
             listeners: {
                 'tabchange': function(tabpanel, tab) {
                     if (!tab) {
-                        /* if the tabpanel has been rendered once and hidden (user 
-                        clicked the remove all button), a tabchange event is triggered 
+                        /* if the tabpanel has been rendered once and hidden (user
+                        clicked the remove all button), a tabchange event is triggered
                         when the tabpanel is displayed again on next query, but
                         the tab is not ready and is yet undefined */
                         return;
@@ -821,7 +839,7 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
                             this.textItem.setText(this.getCount());
                         }
                         // re-enable grid related buttons
-                        this.selectionButton.enable(); 
+                        this.selectionButton.enable();
                         Ext.each(this.selectionActionButton.menu.items.items, function(item) {
                             item.enable();
                         });
@@ -830,7 +848,7 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
                         this.currentGrid = null;
                         this.textItem.setText(this.getCount());
                         // disable grid related buttons
-                        this.selectionButton.disable(); 
+                        this.selectionButton.disable();
                         Ext.each(this.selectionActionButton.menu.items.items, function(item) {
                             item.disable();
                         });
@@ -850,7 +868,7 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
                 scope: this
             },
             bbar: [
-                this.selectionButton, this.selectionActionButton ,'->', 
+                this.selectionButton, this.selectionActionButton, '->',
                 this.messageItem, '-', this.textItem, '-', {
                     text: this.clearAllText,
                     handler: function() {
@@ -864,7 +882,7 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
             ]
         };
 
-        this.tabpan = cgxp.plugins.FeatureGrid.superclass.addOutput.call(this, config);
+        this.tabpan = cgxp.plugins.FeaturesGrid.superclass.addOutput.call(this, config);
         this.tabpan.ownerCt.on("expand", function() {
             this.vectorLayer.setVisibility(true);
         }, this);
@@ -940,4 +958,4 @@ cgxp.plugins.FeatureGrid = Ext.extend(gxp.plugins.Tool, {
     }
 });
 
-Ext.preg(cgxp.plugins.FeatureGrid.prototype.ptype, cgxp.plugins.FeatureGrid);
+Ext.preg(cgxp.plugins.FeaturesGrid.prototype.ptype, cgxp.plugins.FeaturesGrid);