+++ CHANGELOG.rst
@@ -1,5 +1,17 @@
 This file includes migration steps for each release of CGXP.
 
+Version 1.4
+===========
+
+1. Rename ``CGXP/plugins/FeatureGrid.js`` to ``CGXP/plugins/FeaturesGris.js``
+   [typo] (present in ``jsbuild/app.cfg``) and its ``ptype`` from
+   ``cgxp_featuregrid`` to ``cgxp_featuresgrid`` (present in the ``viewer.js``)
+   for name consistency.
+
+2. Add new ``themes`` argument to the plugin ``cgxp_featuresgrid`` to be able to 
+   display the Identifier Attribute on the drawn features.
+
+
 Version 1.3
 ===========
 
@@ -10,6 +22,8 @@ Version 1.3
 2. Former CGXP plugin ``SwitchableWMTSSource`` has been removed and replaced by
    an OpenLayers addin, ``OpenLayers.Layer.SwitchableWMTS``, combined to a 
    standard ``OLSource``. Here is an example of the new syntax to use:
+
+   .. code:: javascript
 
        {
            source: "olsource",
@@ -25,5 +39,4 @@ Version 1.3
                params:  {'time': '200703'},
                group: 'background'
            }, WMTS_OPTIONS)]
-       } 
-
+       }